Activity Log: Update Threat Alerts to v2 of API proposal
As we have discovered more of the details behind threat alerts
we learned that we had to make several UI changes and changes
to the data format to support the task.

In this patch we're updating the visuals and their use of the
underlying data.

This code is still dormant without the `rewind-alerts` flag
enabled and when not in `wp-calypso` or local development.
No data-fetching code exists at this time.